Add P4_16 versions of several example Domino programs
It seems to me that P4_16's addition of C-like assignment syntax has
eliminated any expressiveness advantage that Domino had over P4_14
claimed in the Packet Transactions paper.  All of those advantages
seem to have been of Domino over the more-verbose P4_14, which is
closer to the "triples" form of the compiler mentioned in the paper.